Move action log to common
